Prizes always double

+++ database/migrations/2018_02_19_205338_create_presupuestos_table.php
@@ -23,7 +23,7 @@ class CreatePresupuestosTable extends Migration
             $table->integer('unidades')->default(1);
             $table->string('estado')->default('por comprobar');
             $table->integer('beneficio')->default(0.3);
-            $table->integer('precio_final')->default(0);
+            $table->double('precio_final', 8, 2)->default(0);
             $table->timestamps();
         });
     }
